+++ central/tests/__init__.py
@@ -1,6 +1,0 @@
-import os
-from jaws_central import config
-
-abspath = os.path.abspath(__file__)
-dirpath = os.path.dirname(abspath)
-conf = config.Configuration(os.path.join(dirpath, "..", "jaws-central.ini"))