+++ dashboard/src/store/utility.js
@@ -1,3 +1,0 @@
-export const updateObject = (state, obj) => {
-  return { ...state, ...obj }
-}