+++ jtm/ssh_tunnel_pnnl.sh
@@ -1,1 +1,0 @@
-ssh -L 50000:gwf1.emsl.pnl.gov:15672 suls634@cascade.emsl.pnl.gov