+++ site/jaws_site/perf_metrics.py
@@ -7,8 +7,9 @@ import re
 from typing import Callable
 from pathlib import Path
 from functools import lru_cache
-from jaws_site import runs, tasks
+from jaws_site import runs
 from jaws_rpc import rpc_client_basic
+from jaws_site.cromwell import parse_cromwell_task_dir
 
 logger = logging.getLogger(__package__)
 
@@ -17,7 +18,6 @@ class PerformanceMetrics:
     def __init__(self, session: Callable, rpc_client: rpc_client_basic) -> None:
         self.session = session
         self.rpc_client = rpc_client
-        self.tasks = {}
 
     @lru_cache()
     def get_run_id(self, cromwell_run_id: str) -> int:
@@ -26,8 +26,8 @@ class PerformanceMetrics:
             run = runs.Run.from_cromwell_run_id(self.session, cromwell_run_id)
         except runs.RunNotFoundError or runs.RunDbError as err:
             msg = f"Failed to get run_id from {cromwell_run_id=}: {err}"
-            logger.warn(msg)
-            raise runs.RunDbError(msg)
+            logger.warning(msg)
+            return 0
         return run.data.id
 
     def process_csv(self, csv_file: str) -> list:
@@ -39,7 +39,12 @@ class PerformanceMetrics:
         :return: list of dictionaries where each dictionary is a json doc of the performance metrics.
         :rtype: list
         """
-        csv_data = pd.read_csv(csv_file, parse_dates=[0], index_col=[0])
+        try:
+            csv_data = pd.read_csv(csv_file, parse_dates=[0], index_col=[0])
+        except Exception as err:
+            logging.warning(f"{type(err).__name__} Error opening {csv_file=}")
+            # Return an empty list of dict to be handled later
+            return [{}]
 
         # Remove extranious parts from the current directory
         csv_data["current_dir"] = csv_data.current_dir.apply(remove_beginning_path)
@@ -49,40 +54,37 @@ class PerformanceMetrics:
         # Drops any non-workflow related processes
         csv_data = csv_data[csv_data.cromwell_run_id != "naw"]
 
+        # Removes bash and tee processes which are part of
+        # cromwell or other wrapper scripts
+        csv_data = csv_data[csv_data.name != 'bash']
+        csv_data = csv_data[csv_data.name != 'tee']
+
         # If we have a cromwell-id then get the jaws id by calling jaws db
         csv_data["jaws_run_id"] = csv_data.cromwell_run_id.apply(self.get_run_id)
         # Filter out anything that didn't get a jaws_run_id returned
         csv_data = csv_data[csv_data.jaws_run_id != "None"]
+
+        # Add task_name to the dataframe
+        csv_data["task_name"] = csv_data.current_dir.apply(parse_cromwell_task_dir_name)
 
         # Drops the current_dir, since we shouldn't need it anymore
         # csv_data = csv_data.drop(columns=["current_dir"])
 
         csv_data["@timestamp"] = csv_data.index.map(lambda x: x.isoformat())
         csv_data["mem_total"] = csv_data["mem_rss"] + csv_data["mem_vms"]
+
+        # Group by unique values to get a specific
+        grouped_csv = csv_data.groupby(['cromwell_run_id', 'task_name', 'pid', 'name'])
+        csv_data = grouped_csv.apply(get_grouped_data)
+
+        # Grouping does not preserve the index so we re-set the index to the timestamp
+        csv_data = csv_data.set_index(pd.DatetimeIndex(csv_data['@timestamp']))
+
         csv_data["num_fds"] = csv_data["num_fds"].replace(["None"], np.nan)
         csv_data.fillna(0, inplace=True)
         csv_data["num_fds"] = csv_data.num_fds.astype(int)
 
         return csv_data.to_dict("records")
-
-    @lru_cache()
-    def add_taskname_mapping(self, cromwell_run_id):
-        """Create a cache to map the cromwell task directory to the task name keyed using the cromwell_run_id.
-        The cached result looks something like like:
-
-        self.tasks[some_cromwell_run_id] = {
-            "cromwell-executions/a/aa/aaa": "task_a",
-            "cromwell-executions/b/bb/bbb": "task_b",
-        }
-        """
-        try:
-            self.tasks[cromwell_run_id] = tasks.TaskLog(
-                self.session, cromwell_run_id=cromwell_run_id
-            ).get_task_cromwell_dir_mapping()
-        except Exception:
-            # if calling TaskLog throws an exception (cromwell fails to get data using cromwell_run_id), don't
-            # update the tasks list.
-            pass
 
     def process_metrics(self, done_dir: str, proc_dir: str) -> None:
         """Gather all csv files in done_dir, parse and generate json docs for each entry, add docs to
@@ -111,15 +113,18 @@ class PerformanceMetrics:
             return
 
         for done_file in list(done_dir_obj.glob("*.csv")):
+            logger.info(f"Processing {done_file.name=} ...")
             docs = self.process_csv(done_file)
             for doc in docs:
                 cromwell_run_id = doc.get("cromwell_run_id")
                 run_id = doc.get("jaws_run_id")
+                task_name = doc.get("task_name")
 
                 # Just really a final check at this point since it should already be in the document
                 if not run_id or not cromwell_run_id:
                     logger.error(
-                        f"Error with {run_id=} or {cromwell_run_id=}, Not uploading to performance metrics"
+                        f"Error with {run_id=}, {task_name=}, or {cromwell_run_id=},\
+                                     Not uploading to performance metrics"
                     )
                     continue
 
@@ -127,29 +132,91 @@ class PerformanceMetrics:
                     f"Run {run_id}: Publish performance metrics for cromwell_run_id={cromwell_run_id}"
                 )
 
-                # Add cromwell dir to task name mapping if not already exists
-                self.add_taskname_mapping(cromwell_run_id)
-
-                # Add task name to doc
-                cromwell_dir = doc["current_dir"]
-                if (
-                    cromwell_run_id in self.tasks
-                    and cromwell_dir in self.tasks[cromwell_run_id]
-                ):
-                    task_name = self.tasks[cromwell_run_id][cromwell_dir]
-                else:
-                    task_name = "unknown"
-                doc["task_name"] = task_name
-
                 # Submit doc to RMQ to be picked up by logstash and inserted into elasticsearch
-                self.rpc_client.request(doc)
+                response, status_code = self.rpc_client.request(doc)
+                if status_code:
+                    logger.error(f"Failed to publish metrics into elasticsearch: {response=}")
+                    continue
 
             # Move csv file to processed folder
             processed_file = proc_dir_obj / done_file.name
             Path(f"{done_file}").rename(f"{processed_file}")
 
-        # Clear cached cromwell dir to taskname mapping
-        self.tasks = {}
+
+def compute_rates(data_series: pd.Series, rolling_time: int = 10) -> pd.Series:
+    """Takes a series and computes the derivative over a window and normalizes the values
+        back to a rate or percentage based on the window size.
+
+        :param data_series: Series of data to take the rolling difference on to get rate
+        :type data_series: pd.Series
+        :param rolling_time: Time in seconds for rolling window to look at data over
+        :type rolling_time: int
+        :return: pd.Series
+    """
+    # Gets a rolling window over a specific time period
+    rolling = data_series.rolling(f'{rolling_time}s')
+    # Takes the last - first value to get the difference over the rolling window
+    diff = rolling.apply(lambda x: x.iloc[-1] - x.iloc[0])
+    # Normalize based on the rolling window to get the rate/percentage correclty
+    rate = (100/rolling_time) * diff
+    return rate
+
+
+def get_grouped_data(data: pd.DataFrame, rolling_time: int = 10) -> pd.DataFrame:
+    """Takes a single group from a dataframe which has a groupby
+        and computes cpu and data rates at each piont
+
+        :param data: Dataframe with grouped data
+        :type data: pd.DataFrame
+        :param rolling_time: Time in seconds for rolling window to look at data over
+        :type rolling_time: int
+        :return: pd.DataFrame
+    """
+    # Saves column names from the data
+    col = data.columns
+
+    # Check if there is data in the dataframe
+    if len(data) == 0:
+        logging.warning('No data in group: Empty DataFrame')
+        return pd.DataFrame([], columns=col)
+
+    # Make sure our columns are sorted to compute rates properly
+    data.sort_index(inplace=True)
+    # Gets the start time of the data
+    start_time = data.index[0]
+    # Place the time the processes has been running since at each pont
+    data['running_time'] = (data.index - start_time).total_seconds()
+
+    # remove entries with runtines shorter then the rolling window
+    if np.max(data['running_time']) < rolling_time:
+        # retruns empty dataframe with the correct column names
+        return pd.DataFrame([], columns=col)
+
+    try:
+        # Calculates the percentages for different parts of the cpu per point
+        data['cpu_percentage_user'] = compute_rates(data["cpu_user"], rolling_time=rolling_time)
+        data['cpu_percentage_system'] = compute_rates(data["cpu_system"], rolling_time=rolling_time)
+        data['cpu_percentage_iowait'] = compute_rates(data["cpu_iowait"], rolling_time=rolling_time)
+        data['cpu_percentage_total'] = compute_rates(data["cpu_user"] + data["cpu_system"] +
+                                                     data["cpu_iowait"], rolling_time=rolling_time)
+    except Exception as e:
+        logger.warning(f"Error computing cpu percentage {data=}, {type(e).__name__} : {e}")
+        return pd.DataFrame([], columns=col)
+
+    # remove entries with low cpu usage
+    if np.mean(data['cpu_percentage_total']) < 0.1:
+        # retruns empty dataframe with the correct column names
+        return pd.DataFrame([], columns=col)
+    try:
+        # Calculates the data rates for read and write
+        for rw in ['read', 'write']:
+            # Compute rate and convert from bytes per second to megabytes/s
+            data[f'{rw}_rate_mbs'] = compute_rates(data[f'{rw}_chars']) / (10 * (1024**2))
+    except Exception as e:
+        logger.warning(f"Error computing disk rate {data=}, {type(e).__name__} : {e}")
+        return pd.DataFrame([], columns=col)
+
+    return data
 
 
 @lru_cache()
@@ -168,7 +235,7 @@ def extract_jaws_info(working_dir):
         if match:
             cromwell_run_id = match.group(1)
     except Exception as e:
-        logger.warn(f"Error when processing {working_dir=}, {type(e).__name__} : {e}")
+        logger.warning(f"Error when processing {working_dir=}, {type(e).__name__} : {e}")
 
     return cromwell_run_id
 
@@ -184,6 +251,14 @@ def remove_beginning_path(working_dir):
             # Extract from "cromwell-executions" to the end of the string
             dir_name = working_dir[id_ce_dir:]
     except Exception as e:
-        logger.warn(f"Error when processing {working_dir=}, {type(e).__name__} : {e}")
+        logger.warning(f"Error when processing {working_dir=}, {type(e).__name__} : {e}")
 
     return dir_name
+
+
+@lru_cache()
+def parse_cromwell_task_dir_name(results):
+    """Get's just the name from parse_cromwell_task_dir
+    Needed for pandas apply since we just want the name in that column
+    """
+    return parse_cromwell_task_dir(results)['name']